Refactored nestedPage to follow other prosemirror node convention
