+++ components/common/CharmEditor/InlineCharmEditor.tsx
@@ -8,7 +8,7 @@ import {
   underline
 } from '@bangle.dev/base-components';
 import debounce from 'lodash/debounce';
-import { NodeView, Plugin, SpecRegistry } from '@bangle.dev/core';
+import { Plugin, SpecRegistry } from '@bangle.dev/core';
 import { EditorView, Node, PluginKey } from '@bangle.dev/pm';
 import { useEditorState } from '@bangle.dev/react';
 import { CSSProperties, ReactNode, useState } from 'react';
@@ -17,7 +17,7 @@ import { BangleEditor as ReactBangleEditor } from 'components/common/CharmEditor
 import { PageContent } from 'models';
 import FloatingMenu, { floatingMenuPlugin } from './components/FloatingMenu';
 import EmojiSuggest, { plugins as emojiPlugins, specs as emojiSpecs } from './components/emojiSuggest';
-import { MentionNode, mentionPlugins, mentionSpecs, MentionSuggest } from './components/mention';
+import Mention, { mentionPlugins, mentionSpecs, MentionSuggest, mentionPluginKeyName } from './components/mention';
 import * as tabIndent from './components/tabIndent';
 import Placeholder from './components/Placeholder';
 import { checkForEmpty } from './utils';
@@ -27,8 +27,8 @@ export interface ICharmEditorOutput {
   rawText: string
 }
 
-const emojiSuggestPluginKey = new PluginKey('emojiSuggest');
-const mentionSuggestPluginKey = new PluginKey('mentionSuggest');
+const emojiPluginKey = new PluginKey('emojiSuggest');
+const mentionPluginKey = new PluginKey(mentionPluginKeyName);
 const floatingMenuPluginKey = new PluginKey('floatingMenu');
 
 export const specRegistry = new SpecRegistry([
@@ -72,10 +72,10 @@ export function charmEditorPlugins (
     strike.plugins(),
     underline.plugins(),
     emojiPlugins({
-      key: emojiSuggestPluginKey
+      key: emojiPluginKey
     }),
     mentionPlugins({
-      key: mentionSuggestPluginKey
+      key: mentionPluginKey
     }),
     floatingMenuPlugin({
       key: floatingMenuPluginKey,
@@ -185,9 +185,9 @@ export default function CharmEditor (
         switch (props.node.type.name) {
           case 'mention': {
             return (
-              <MentionNode {...props}>
+              <Mention {...props}>
                 {_children}
-              </MentionNode>
+              </Mention>
             );
           }
           default: {
@@ -197,8 +197,8 @@ export default function CharmEditor (
       }}
     >
       <FloatingMenu inline pluginKey={floatingMenuPluginKey} />
-      <MentionSuggest pluginKey={mentionSuggestPluginKey} />
-      <EmojiSuggest pluginKey={emojiSuggestPluginKey} />
+      <MentionSuggest pluginKey={mentionPluginKey} />
+      <EmojiSuggest pluginKey={emojiPluginKey} />
       {children}
     </StyledReactBangleEditor>
   );
