+++ components/common/CharmEditor/components/mention/index.ts
@@ -1,6 +1,11 @@
-export * from './components';
+import { Mention, MentionSuggest } from './components';
+
 export * from './mention.constants';
 export * from './mention.plugins';
 export * from './mention.interfaces';
 export * from './mention.specs';
 export * from './mention.utils';
+export {
+  MentionSuggest
+};
+export default Mention;
