+++ components/common/CharmEditor/components/nestedPage/components/NestedPagesList.tsx
@@ -7,9 +7,9 @@ import { PluginKey } from 'prosemirror-state';
 import { safeScrollIntoViewIfNeeded } from 'lib/browser';
 import { insertNestedPage } from 'lib/prosemirror/insertNestedPage';
 import { hideSuggestionsTooltip, SuggestTooltipPluginState } from '../../@bangle.dev/tooltip/suggest-tooltip';
-import { NestedPagePluginState } from '../nestedPage';
 import PopoverMenu, { GroupLabel } from '../../PopoverMenu';
 import { PagesList } from '../../PageList';
+import { NestedPagePluginState } from '../nestedPage.interfaces';
 
 function NestedPagesList ({ pluginKey }: {pluginKey: PluginKey<NestedPagePluginState>}) {
   const { pages } = usePages();
