+++ components/common/CharmEditor/components/nestedPage/index.ts
@@ -1,7 +1,12 @@
 import NestedPage from './components/NestedPage';
 import NestedPagesList from './components/NestedPagesList';
 
-export * from './nestedPage';
+export * from './nestedPage.constants';
+export * from './nestedPage.interfaces';
+export * from './nestedPage.plugins';
+export * from './nestedPage.specs';
+export * from './nestedPage.utils';
+
 export { NestedPagesList };
 
 export default NestedPage;
