+++ components/common/CharmEditor/components/nestedPage/nestedPage.constants.ts
@@ -1,0 +1,4 @@
+export const nestedPageNodeName = 'page';
+export const nestedPageSuggestMarkName = 'nestedPageSuggest';
+export const nestedPageMarkdownEnclosure = 'MARKDOWN_NESTED_PAGE';
+export const nestedPagePluginKeyName = 'nestedPage';
