+++ components/common/CharmEditor/components/nestedPage/nestedPage.interfaces.ts
@@ -1,0 +1,4 @@
+import { SuggestPluginState } from 'lib/prosemirror/interfaces';
+
+export interface NestedPagePluginState extends SuggestPluginState{
+}
