+++ components/common/CharmEditor/components/nestedPage/nestedPage.specs.ts
@@ -1,0 +1,38 @@
+import { RawSpecs } from '@bangle.dev/core';
+import { DOMOutputSpec } from '@bangle.dev/pm';
+import { nestedPageNodeName, nestedPageSuggestMarkName } from './nestedPage.constants';
+import * as suggestTooltip from '../@bangle.dev/tooltip/suggest-tooltip';
+import { encloseNestedPage } from './nestedPage.utils';
+
+export function nestedPageSpec (): RawSpecs {
+  return [{
+    type: 'node',
+    name: nestedPageNodeName,
+    schema: {
+      inline: false,
+      attrs: {
+        // This property is used to reference the page
+        id: {
+          default: null
+        }
+      },
+      group: 'block',
+      draggable: false,
+      parseDOM: [{ tag: 'div.charm-nested-page' }],
+      toDOM: (): DOMOutputSpec => {
+        return ['div', { class: 'charm-nested-page' }];
+      }
+    },
+    markdown: {
+      toMarkdown: (state, node) => {
+        try {
+          state.write(encloseNestedPage(node.attrs.id));
+          state.ensureNewLine();
+        }
+        catch (err) {
+          console.log('Conversion err', err);
+        }
+      }
+    }
+  }, suggestTooltip.spec({ markName: nestedPageSuggestMarkName })];
+}
